UPBGE: Adapt again txt_to_buf to use size_t lenght

+++ source/blender/python/generic/bpy_internal_import.c
@@ -119,7 +119,7 @@ bool bpy_text_compile(Text *text)
 
   fn_dummy_py = PyC_UnicodeFromByte(fn_dummy);
 
-  int buf_len_dummy;
+  size_t buf_len_dummy;
   buf = txt_to_buf(text, &buf_len_dummy);
   text->compiled = Py_CompileStringObject(buf, fn_dummy_py, Py_file_input, NULL, -1);
   MEM_freeN(buf);
